+++ jabgui/src/main/java/org/jabref/gui/actions/StandardActions.java
@@ -38,6 +38,7 @@ public enum StandardActions implements Action {
     OPEN_URL(Localization.lang("Open URL or DOI"), IconTheme.JabRefIcons.WWW, KeyBinding.OPEN_URL_OR_DOI),
     SEARCH_SHORTSCIENCE(Localization.lang("Search ShortScience")),
     SEARCH_GOOGLE_SCHOLAR(Localization.lang("Search Google Scholar")),
+    SEARCH_SEMANTIC_SCHOLAR(Localization.lang("Search Semantic Scholar")),
     SEARCH(Localization.lang("Search...")),
     MERGE_WITH_FETCHED_ENTRY(Localization.lang("Get bibliographic data from %0", "DOI/ISBN/..."), KeyBinding.MERGE_WITH_FETCHED_ENTRY),
     BATCH_MERGE_WITH_FETCHED_ENTRY(Localization.lang("Get bibliographic data from %0 (fully automated)", "DOI/ISBN/...")),